#74 Campe作成ページを編集機能にも対応
