+++ lib/view/campe_tile.dart
@@ -1,6 +1,6 @@
 import 'package:campe_app/controller/campe_list_controller/campe_lsit_controller_provider.dart';
 import 'package:campe_app/view/screens/home_screen.dart';
-import 'package:campe_app/view/screens/update_screen.dart';
+import 'package:campe_app/view/screens/making_screen.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
@@ -14,7 +14,7 @@ class CampeTile extends ConsumerWidget {
       title: Text(campe.name),
       onTap: () {
         Navigator.push(context, MaterialPageRoute(builder: (context) {
-          return UpdateScreen(oldText: campe.name);
+          return  MakingScreen(campe: campe);
         }));
       },
       onLongPress: () {
