+++ pubspec.yaml
@@ -26,6 +26,8 @@ dependencies:
   flutter_native_splash: ^1.3.1
   vibration: ^1.7.4-nullsafety.0
   flutter_easyloading: ^3.0.3
+  flutter_hooks: ^0.18.2+1
+  hooks_riverpod: ^1.0.3
 
 
 dev_dependencies:
